Remove akka-slf4j
It is not needed to specify the dependency explicitly since we use akka-actor-typed instead of akka-actor.
akka-actor-typed already depends on akka-slf4.

We can check the above dependencies from the below links.
https://search.maven.org/artifact/com.typesafe.akka/akka-actor-typed_2.13/2.6.12/jar
https://search.maven.org/artifact/com.typesafe.akka/akka-actor_2.13/2.6.12/jar

We can ensure that akka-actor requires us to add akka-slf4j explicitly from the following link.
https://doc.akka.io/docs/akka/current/logging.html#slf4j

Also, we can ensure that akka-actor-typed does not from the following link.
https://doc.akka.io/docs/akka/current/typed/logging.html#introduction

The dependency list contains akka-slf4j even if we remove the direct dependency.
```
$ sbt dependencyList | grep akka-slf4j | sort | uniq
[info] com.typesafe.akka:akka-slf4j_2.13:2.6.12
```